Add the concept of a HitJudgementResolver.
