--- osu.Game.Mode.Osu/ComboJudgement.cs
@@ -1,14 +1,0 @@
-﻿using System.ComponentModel;
-
-namespace osu.Game.Modes.Osu
-{
-    public enum ComboJudgement
-    {
-        [Description(@"")]
-        None,
-        [Description(@"Good")]
-        Good,
-        [Description(@"Amazing")]
-        Perfect
-    }
-}
