+++ osu.Game.Mode.Osu/osu.Game.Modes.Osu.csproj
@@ -43,13 +43,11 @@
   <ItemGroup>
     <Compile Include="Objects\Drawables\Pieces\ApproachCircle.cs" />
     <Compile Include="Objects\Drawables\Pieces\CirclePiece.cs" />
-    <Compile Include="ComboJudgement.cs" />
     <Compile Include="Objects\Drawables\DrawableSlider.cs" />
     <Compile Include="Objects\Drawables\Pieces\ExplodePiece.cs" />
     <Compile Include="Objects\Drawables\Pieces\FlashPiece.cs" />
     <Compile Include="Objects\Drawables\Pieces\GlowPiece.cs" />
     <Compile Include="Objects\Drawables\HitExplosion.cs" />
-    <Compile Include="Judgement.cs" />
     <Compile Include="Objects\Drawables\Pieces\NumberPiece.cs" />
     <Compile Include="Objects\Drawables\Pieces\RingPiece.cs" />
     <Compile Include="Objects\Drawables\Pieces\Triangles.cs" />
